Add files via upload
This is the first challenge for our circuit python. The question is:
Create a circuit with a potentiometer and 2 LEDs.
As the Potentiometer is turned up the first LED gets brighter.  Once the Potentiometer is at 50% the first LED is at full brightness. Continue turning the potentiometer and the second LED gets brighter.
There is no need for a transistor in this circuit
Pay attention to your choice of resistor.